// src/app/model/enquirycampaign.ts
export interface EnquiryCampaign {
    institute_enquiry_id: any;
    institution_id: any;
    enquiry_no: any;
    name: any;
    phone: any;
    email: any;
    standard: any;
    subjects: any;
    status: any;
    statusValue: any;
    is_converted: any;
    followUpDate: any;
    occupation_id: any;
    school_id: any;
    enquiry_date: any;
    standard_id: any;
    subject_id: any;
    referred_by: any;
    source_id: any;
    priority: any;
    follow_type: any;
    assigned_name: any;
    is_recent: any;
    slot_id: any;
    slot: any;
    updateDate: any;
    isDashbord: any;
    isRport: any;
    totalcount: any;
    newEnqcount: any;
    enquiry_no_date: any;
    name_person: any;
    followUpDateTime: any;
    standard_subject: any;
    closedReasonText: any;
    followUpTime: any;
    filtered_statuses: any;
    filtered_slots: any;
} 